+++ android/local.properties
@@ -1,5 +1,5 @@
-flutter.sdk=C:\\Users\\santi\\Flutter\\flutter
-sdk.dir=C:\\Users\\santi\\AppData\\Local\\Android\\sdk
+flutter.sdk=C:\\flutter\\flutter
+sdk.dir=C:\\Users\\olvei\\AppData\\Local\\Android\\sdk
 flutter.buildMode=debug
 flutter.versionName=1.0.0
 flutter.versionCode=1
