+++ lib/GUI/Pages/loginScreen.dart
@@ -1,0 +1,79 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/foundation.dart';
+import 'homeScreen.dart';
+
+//todavia le falta a esta screen, adicionalmente de la screen de registro
+class LogInPage extends StatefulWidget {
+  const LogInPage({super.key});
+
+  @override
+  LogInPageState createState() => LogInPageState();
+}
+
+class LogInPageState extends State<LogInPage> {
+  final _usernameController = TextEditingController();
+  final _passwordController = TextEditingController();
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: Theme.of(context).colorScheme.surface,
+      body: Center(
+        child: Padding(
+          padding: const EdgeInsets.all(16.0),
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              Text(
+                'Inicio de sesión',
+                style: Theme.of(context).textTheme.titleMedium,
+              ),
+              const SizedBox(height: 20),
+
+              // Username input field
+              TextField(
+                controller: _usernameController,
+                decoration: InputDecoration(
+                  labelText: 'Correo',
+                  labelStyle: TextStyle(color: Colors.grey),
+                  border: OutlineInputBorder(),
+                  filled: true,
+                  fillColor: Colors.white,
+                ),
+                style: TextStyle(color: Colors.black),
+              ),
+              const SizedBox(height: 20),
+
+              // Password input field
+              TextField(
+                controller: _passwordController,
+                obscureText: true,
+                decoration: InputDecoration(
+                  labelText: 'Contraseña',
+                  labelStyle: TextStyle(color: Colors.grey),
+                  border: OutlineInputBorder(),
+                  filled: true,
+                  fillColor: Colors.white,
+                ),
+                style: TextStyle(color: Colors.black),
+              ),
+              const SizedBox(height: 30),
+
+              // Login button
+              ElevatedButton(
+                onPressed: () {
+                  // acá faltan chequeos y guardado de informacion del login
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(builder: (context) => const HomeScreen()),
+                  );
+                },
+                child: const Text('Iniciar sesión'),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
