+++ lib/GUI/Pages/profileScreen.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class ProfilePage extends StatelessWidget {
+  const ProfilePage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Text(
+        'Página Perfil',
+        style: Theme.of(context).textTheme.titleMedium,
+      ),
+    );
+  }
+}
