+++ lib/GUI/theme/dark_mode.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+ThemeData darkMode = ThemeData(
+  brightness: Brightness.dark,
+  colorScheme: ColorScheme.dark(
+    background: Colors.grey.shade900,
+    primary: Colors.grey.shade800,
+    secondary: Colors.grey.shade700,
+    inversePrimary: Colors.grey.shade300,
+  ),
+  textTheme: ThemeData.dark().textTheme.apply(
+    bodyColor: Colors.grey[300],
+    displayColor: Colors.white,
+
+  ),
+);
