+++ lib/main.dart
@@ -1,8 +1,8 @@
 // ignore_for_file: deprecated_member_use, library_private_types_in_public_api, use_key_in_widget_constructors
 
 import 'package:flutter/material.dart';
-import 'package:flutter_prueba/MongoManager/Constant.dart';
-import 'package:flutter_prueba/Schedules/Schedule.dart';
+import 'package:metrosync/MongoManager/Constant.dart';
+import 'package:metrosync/Schedules/Schedule.dart';
 import 'package:mongo_dart/mongo_dart.dart'
     show Db, DbCollection, where, modify;
 import 'MongoManager/MongoDB.dart';
