+++ test/widget_test.dart
@@ -7,8 +7,9 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:metrosync/main.dart';
+import 'package:metrosync/GUI/GUI.dart';
 
-import 'package:flutter_prueba/main.dart';
 
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
